backend/requirements.txt
Flask>=2.2.5
Flask-Cors>=3.0.10
Werkzeug>=2.2.3
gunicorn>=20.1.0
mysql-connector-python>=8.4.0
python-dotenv>=1.0.0